backend/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
executing==2.0.1
filetype==1.2.0
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
isort==5.13.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
urllib3==2.1.0
wcwidth==0.2.13